Internal change
PiperOrigin-RevId: 283770513
